BOOM.
I fixed you, you son of a gun.

I know it’s not right for the purpose of this assignment, but I added
the <p> element to make the picture of Venus in the right spot. It was
in the middle of the page, right next to the line of text so I used the
<p> element to make that line of text it’s own thing. Commence the dock
of marks for my awful semi detailed commits.